requirements.txt
tensorflow
keras
numpy
pandas
matplotlib
seaborn
scikit-learn
opencv-python
Pillow
streamlit
torch
torchvision
tqdm
gdown

